+++ src/main/java/com/dnd/gongmuin/security/oauth2/AuthInfo.java
@@ -1,4 +1,4 @@
-package com.dnd.gongmuin.auth.dto;
+package com.dnd.gongmuin.security.oauth2;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,25 +6,25 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class AuthDto {
+public class AuthInfo {
 
 	private String socialName;
 	private String socialEmail;
 
 	@Builder
-	private AuthDto(String socialName, String socialEmail) {
+	private AuthInfo(String socialName, String socialEmail) {
 		this.socialName = socialName;
 		this.socialEmail = socialEmail;
 	}
 
-	public static AuthDto fromSocialEmail(String socialEmail) {
-		return AuthDto.builder()
+	public static AuthInfo fromSocialEmail(String socialEmail) {
+		return AuthInfo.builder()
 			.socialEmail(socialEmail)
 			.build();
 	}
 
-	public static AuthDto of(String socialName, String socialEmail) {
-		return new AuthDto(socialName, socialEmail);
+	public static AuthInfo of(String socialName, String socialEmail) {
+		return new AuthInfo(socialName, socialEmail);
 	}
 
 }
